Actually logging docker registry HTTP
